# smart-api/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  description: >-
    An project that has how the main functionality recover packings on the
    system through Beacons and Checkpoints
  title: Reciclapac API Documentation
  contact:
    name: David W. F. Santos
    email: david.santos@pe.senai.br
    url: 'http://pe.senai.br'
host: 'localhost:3000'
basePath: /api
schemes:
  - http
  - https
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
...
securityDefinitions:
  Bearer:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  '/inventory/logistic/general/{limit}/{page}':
    x-swagger-router-controller: inventory_logistic
    post:
      tags:
        - Inventory Logistic
      summary: >-
        Retrieve all information about equipaments on the system related to
        logistic operator with pagination and without some filter. Groups
        equipaments with the same code, supplier and project to generate
        inventory
      description: ''
      operationId: general_inventory_packing
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of packings that will be Retrieved
          required: true
          type: string
        - in: path
          name: page
          description: Retrieve inventory related this page
          required: true
          type: string
        - in: body
          name: body
          description: >-
            Array of the suppliers related to the logistic operator that will be
            used to Retrieve the information about the packings
          required: true
          schema:
            $ref: '#/definitions/ArraySupplier'
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/inventory/logistic/permanence/{limit}/{page}/{code}':
    x-swagger-router-controller: inventory_logistic
    post:
      tags:
        - Inventory Logistic
      summary: >-
        Retrieve all information about permanence of the equipaments on the
        system related to logistic operator with pagination and filters. Groups
        equipaments with the same code, supplier and project to generate
        inventory
      description: ''
      operationId: inventory_permanence
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of packings that will be Retrieveed
          required: true
          type: string
        - in: path
          name: page
          description: Retrieve inventory related this page
          required: true
          type: string
        - in: path
          name: code
          description: Retrieve inventory related this packing code
          required: true
          type: string
        - in: body
          name: body
          description: >-
            Array of the suppliers related to the logistic operator that will be
            used to Retrieve the information about the packings
          required: true
          schema:
            $ref: '#/definitions/ArraySupplier'
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/inventory/logistic/quantity/{limit}/{page}/{code}':
    x-swagger-router-controller: inventory_logistic
    post:
      tags:
        - Inventory Logistic
      summary: >-
        Retrieve all information about quantity of the equipaments  on the
        system related to logistic operator with pagination and filters. Groups
        equipaments with the same code, supplier and project to generate
        inventory
      description: ''
      operationId: quantity_inventory
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of packings that will be Retrieveed
          required: true
          type: string
        - in: path
          name: page
          description: Retrieve inventory related this page
          required: true
          type: string
        - in: path
          name: code
          description: Retrieve inventory related this packing code
          required: true
          type: string
        - in: body
          name: body
          description: >-
            Array of the suppliers related to the logistic operator that will be
            used to Retrieve the information about the packings
          required: true
          schema:
            $ref: '#/definitions/ArraySupplier'
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/inventory/logistic/packing/historic/{limit}/{page}/{serial}/{code}':
    x-swagger-router-controller: inventory_logistic
    post:
      tags:
        - Inventory Logistic
      summary: >-
        Retrieve all information about equipaments hisotric  on the system
        related to logistic operator with pagination and filters. Groups
        equipaments with the same code, supplier and project to generate
        inventory
      description: ''
      operationId: inventory_packing_historic
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of packings that will be Retrieveed
          required: true
          type: string
        - in: path
          name: page
          description: Retrieve inventory related this page
          required: true
          type: string
        - in: path
          name: serial
          description: Retrieve inventory related this packing serial
          required: true
          type: string
        - in: path
          name: code
          description: Retrieve inventory related this packing code
          required: true
          type: string
        - in: body
          name: body
          description: >-
            Array of the suppliers related to the logistic operator that will be
            used to Retrieve the information about the packings
          required: true
          schema:
            $ref: '#/definitions/ArraySupplier'
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/inventory/logistic/battery/{limit}/{page}':
    x-swagger-router-controller: inventory_logistic
    post:
      tags:
        - Inventory Logistic
      summary: >-
        Retrieve all information about battery level of the equipaments on the
        system related to logistic operator  with pagination and filters. Groups
        equipaments with the same code, supplier and project to generate
        inventory
      description: ''
      operationId: inventory_battery
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of packings that will be Retrieveed
          required: true
          type: string
        - in: path
          name: page
          description: Retrieve inventory related this page
          required: true
          type: string
        - in: query
          name: code
          description: Retrieve inventory related this packing code
          required: false
          type: string
        - in: body
          name: body
          description: >-
            Array of the suppliers related to the logistic operator that will be
            used to Retrieve the information about the packings
          required: true
          schema:
            $ref: '#/definitions/ArraySupplier'
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/inventory/logistic/general/packings/{limit}/{page}':
    x-swagger-router-controller: inventory_logistic
    post:
      tags:
        - Inventory Logistic
      summary: >-
        Retrieve all information about historic equipaments on the system
        related to logistic operator with pagination without someone filter.
        Groups equipaments with the same code, supplier and project to generate
        inventory
      description: ''
      operationId: inventory_packings
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of packings that will be Retrieveed
          required: true
          type: string
        - in: path
          name: page
          description: Retrieve inventory related this page
          required: true
          type: string
        - in: query
          name: code
          description: Retrieve inventory related this packing code
          required: false
          type: string
        - in: body
          name: body
          description: >-
            Array of the suppliers related to the logistic operator that will be
            used to Retrieve the information about the packings
          required: true
          schema:
            $ref: '#/definitions/ArraySupplier'
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/inventory/general/{limit}/{page}':
    x-swagger-router-controller: packing_controller
    get:
      tags:
        - Inventory
      summary: >-
        Retrieve all information about equipaments on the system with pagination
        and without some filter. Groups equipaments with the same code, supplier
        and project to generate inventory
      description: ''
      operationId: general_inventory_packing
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of information about packings that will be returned
          required: true
          type: string
        - in: path
          name: page
          description: Return inventory related this page
          required: true
          type: string
        - in: query
          name: attr
          description: Return inventory related this supplier
          required: false
          type: string
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/inventory/permanence/{limit}/{page}/{code}':
    x-swagger-router-controller: packing_controller
    get:
      tags:
        - Inventory
      summary: >-
        Retrieve all information about permancence of the equipaments on the
        system with pagination and some filters. Groups equipaments with the
        same code, supplier and project to generate inventory
      description: ''
      operationId: inventory_permanence
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of packings that will be returned
          required: true
          type: string
        - in: path
          name: page
          description: Retrieve inventory related this page
          required: true
          type: string
        - in: path
          name: code
          description: Retrieve inventory related this packing code
          required: true
          type: string
        - in: query
          name: attr
          description: Retrieve inventory related this supplier id related to packing
          required: false
          type: string
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/inventory/quantity/{limit}/{page}/{code}':
    x-swagger-router-controller: packing_controller
    get:
      tags:
        - Inventory
      summary: >-
        Retrieve all information about quantity of the equipaments on the system
        with pagination and some filters. Groups equipaments with the same code,
        supplier and project to generate inventory
      description: ''
      operationId: quantity_inventory
      security:
        - Bearer: []
      parameters:
        - in: path
          name: limit
          description: The limit of packings that will be returned
          required: true
          type: string
        - in: path
          name: page
          description: Return inventory related this page
          required: true
          type: string
        - in: path
          name: code
          description: Return inventory related this packing code
          required: true
          type: string
        - in: query
          name: attr
          description: Retrieve inventory related this supplier id related to packing
          required: false
          type: string
      responses:
        '200':
          description: Successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          description: Client Error
          schema:
            $ref: '#/definitions/Error400'
        '500':
          description: Server Error
          schema:
            $ref: '#/definitions/Error500'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
  ErrorResponse:
    type: object
    properties:
      jsonapi:
        type: object
        description: a API version
      name:
        type: string
        description: error type
      message:
        type: string
        description: a short comment about the error
      error:
        type: object
        description: details about the error
      details:
        type: string
        description: especific details about the error
    required:
      - jsonapi
  Error400:
    type: object
    properties:
      jsonapi:
        type: object
        description: a API version
      name:
        type: string
        description: error type
      message:
        type: string
        description: a short comment about the error
      error:
        type: object
        description: details about the error
      details:
        type: string
        description: especific details about the error
    required:
      - jsonapi
  Error500:
    type: object
    properties:
      jsonapi:
        type: object
        description: a API version
      name:
        type: string
        description: error type
      message:
        type: string
        description: a short comment about the error
      error:
        type: object
        description: details about the error
      details:
        type: string
        description: especific details about the error
    required:
      - jsonapi
...
  SuccessResponse:
    type: object
    properties:
      jsonapi:
        type: object
        description: a API version
    required:
      - jsonapi
...
  ArraySupplier:
    type: array
    items:
      type: string
...